pyproject.toml
# Please list the required python packages.

[project]
name = "matsuzawa-20260227"
version = "0.1.0"
description = "Prompt-only Chain-of-Thought evaluation experiment"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "datasets>=2.16.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "pydantic>=2.6.0",
    "numpy>=1.24.0",
    "matplotlib>=3.8.0",
    "scipy>=1.11.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
